# api/openapi.yaml
openapi: 3.0.3
info:
  title: go-mumble-server REST API
  version: 0.1.0
  description: Management API for the Mumble voice chat server
components:
  schemas:
    ConnectedUser:
      type: object
      description: A connected Mumble user (for channel tree display)
      properties:
        session_id:
          type: integer
          format: uint32
          description: Mumble session ID
        user_id:
          type: integer
          format: uint32
          description: Mumble user ID (0 = SuperUser; 0x80000000+ = API user)
        name:
          type: string
          description: Display name
        channel_id:
          type: integer
          format: uint32
          description: Channel the user is in
        address:
          type: string
          description: Client IP address (host, no port)
        ping:
          type: number
          format: float
          description: Client-reported TCP ping in milliseconds (from last Ping message)
        certificate_hash:
          type: string
          description: SHA-1 hex fingerprint of the client's TLS certificate (lowercase); empty if client did not present a certificate. Use for banning by certificate.
        self_mute:
          type: boolean
          description: User has muted themselves
        self_deaf:
          type: boolean
          description: User has deafened themselves
        mute:
          type: boolean
          description: Server has muted the user
        deaf:
          type: boolean
          description: Server has deafened the user
        is_admin:
          type: boolean
          description: User has admin privileges (SuperUser, API role=admin, or in @admin group)
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /api/v1/auth/login
paths:
...
  /api/v1/auth/login:
    post:
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Returns token and user
  /api/v1/auth/register:
    post:
      summary: Register
      description: First user becomes admin
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: User created, returns token and user
  /api/v1/users:
    get:
      summary: List users (admin)
      operationId: listUsersAdmin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      summary: Update user role (admin)
      operationId: updateUserRole
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, user] }
      responses:
        '204':
          description: Updated
    delete:
      summary: Delete user (admin)
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
...
  /api/v1/servers:
    get:
      summary: List virtual servers
      operationId: listServers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of virtual servers
    post:
      summary: Create virtual server
      operationId: createServer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                host: { type: string }
                port: { type: integer }
                max_users: { type: integer }
                welcome_text: { type: string }
      responses:
        '201':
          description: Created
        '400':
          description: Invalid request
  /api/v1/servers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get virtual server
      operationId: getServer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Virtual server
        '404':
          description: Not found
    patch:
      summary: Update virtual server
      operationId: updateServer
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                host: { type: string }
                port: { type: integer }
                max_users: { type: integer }
                welcome_text: { type: string }
      responses:
        '200':
          description: Updated
        '404':
          description: Not found
    delete:
      summary: Delete virtual server
      operationId: deleteServer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /api/v1/servers/{id}/bans:
    get:
      summary: List bans
      operationId: listBans
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of bans
    post:
      summary: Add ban
      operationId: createBan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                address: { type: string }
                hash: { type: string }
                name: { type: string }
                reason: { type: string }
                duration: { type: integer }
      responses:
        '201':
          description: Created
  /api/v1/servers/{id}/bans/{banId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: banId
        in: path
        required: true
        schema:
          type: integer
    delete:
      summary: Remove ban
      operationId: deleteBan
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /api/v1/servers/{id}/users:
    get:
      summary: List connected users
      description: Returns Mumble clients connected to this virtual server, with is_admin for RBAC display
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of connected users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectedUser'
  /api/v1/servers/{id}/users/{sessionId}/kick:
    post:
      summary: Kick connected user
      operationId: kickUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '204':
          description: Kicked
        '404':
          description: User not found
  /api/v1/servers/{id}/users/{sessionId}/mute:
    post:
      summary: Mute or unmute connected user
      operationId: muteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [mute]
              properties:
                mute: { type: boolean }
      responses:
        '204':
          description: Updated
        '404':
          description: User not found
  /api/v1/servers/{id}/users/{sessionId}/ban:
    post:
      summary: Ban and kick connected user
      operationId: banUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '204':
          description: Banned and kicked
        '404':
          description: User not found
  /api/v1/servers/{id}/channels:
    get:
      summary: Get channel tree
      operationId: getChannels
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Channel tree
    post:
      summary: Create channel
      operationId: createChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                parent_id: { type: integer }
                name: { type: string }
                description: { type: string }
                position: { type: integer }
                is_temporary: { type: boolean }
                max_users: { type: integer }
      responses:
        '201':
          description: Created
  /api/v1/servers/{id}/channels/{channelId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: channelId
        in: path
        required: true
        schema:
          type: integer
    patch:
      summary: Update channel
      operationId: updateChannel
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                position: { type: integer }
                max_users: { type: integer }
                is_temporary: { type: boolean }
                links: { type: array, items: { type: integer } }
      responses:
        '204':
          description: Updated
    delete:
      summary: Delete channel
      operationId: deleteChannel
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
  /api/v1/servers/{id}/channels/{channelId}/acl:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: channelId
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get channel ACL and groups
      operationId: getChannelACL
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ACL and groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        name: { type: string }
                        inherit: { type: boolean }
                        inheritable: { type: boolean }
                        add_user_ids: { type: array, items: { type: integer } }
                        remove_user_ids: { type: array, items: { type: integer } }
                  acls:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        priority: { type: integer }
                        apply_here: { type: boolean }
                        apply_subs: { type: boolean }
                        user_id: { type: integer, nullable: true }
                        group_name: { type: string }
                        access_token: { type: string }
                        eval_here: { type: boolean, description: "~ prefix: resolve in ACL-definition channel context" }
                        invert: { type: boolean, description: "! prefix: invert selector" }
                        grant: { type: integer, description: "Permission bitmask" }
                        deny: { type: integer, description: "Permission bitmask" }
    put:
      summary: Replace channel ACL and groups
      operationId: putChannelACL
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [groups, acls]
              properties:
                groups:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      inherit: { type: boolean }
                      inheritable: { type: boolean }
                      add_user_ids: { type: array, items: { type: integer } }
                      remove_user_ids: { type: array, items: { type: integer } }
                acls:
                  type: array
                  items:
                    type: object
                    properties:
                      priority: { type: integer }
                      apply_here: { type: boolean }
                      apply_subs: { type: boolean }
                      user_id: { type: integer, nullable: true }
                      group_name: { type: string }
                      access_token: { type: string }
                      eval_here: { type: boolean }
                      invert: { type: boolean }
                      grant: { type: integer }
                      deny: { type: integer }
      responses:
        '204':
          description: Updated